+++ routes/web.php
@@ -1,9 +1,10 @@
 <?php
 
-use App\Http\Controllers\ListingController;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
 /*
 |--------------------------------------------------------------------------
@@ -16,7 +17,7 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/user/{id}', function ($id){
+Route::get('/user/{id}', function ($id) {
     return response("USER ID IS:" . $id);
 });
 
@@ -25,22 +26,41 @@ Route::get('/', [ListingController::class, 'index']);
 
 
 // Show Create Form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store Listing Data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show Edit From
 
-Route::get('listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update Lisitng Item
 
-Route::put('/listings/{listing}', [ListingController::class , 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete Lisitng Item
 
-Route::delete('/listings/{listing}', [ListingController::class , 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // Get Single Listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
+
+// Show Register From
+
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+// Create New User
+Route::post('users', [UserController::class, 'store']);
+
+// Logout
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Login
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
